+++ src/gemseo_fmu/disciplines/fmu_discipline.py
@@ -18,9 +18,17 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
+from gemseo.datasets.dataset import Dataset
+from gemseo.post.dataset.lines import Lines
+from numpy import newaxis
+
 from gemseo_fmu.disciplines.base_fmu_discipline import BaseFMUDiscipline
+from gemseo_fmu.utils.time_duration import TimeDuration
 
 if TYPE_CHECKING:
+    from collections.abc import Iterable
+    from pathlib import Path
+
     from numpy.typing import NDArray
 
 
@@ -40,6 +48,8 @@ class FMUDiscipline(BaseFMUDiscipline):
         the time series are considered as constant.
     """
 
+    TimeUnit = TimeDuration.TimeUnit
+
     @property
     def initial_values(self) -> dict[str, NDArray[float]]:
         """The initial input, output and time values."""
@@ -49,3 +59,47 @@ class FMUDiscipline(BaseFMUDiscipline):
     def time(self) -> NDArray[float] | None:
         """The time steps of the last execution if any."""
         return self._time
+
+    def plot(
+        self,
+        output_names: str | Iterable[str],
+        abscissa_name: str = "",
+        time_unit: TimeUnit = TimeUnit.SECONDS,
+        save: bool = True,
+        show: bool = False,
+        file_path: str | Path = "",
+    ) -> Lines:
+        """Plot the time evolution of output variables.
+
+        Args:
+            output_names: The name(s) of the output variable(s).
+            abscissa_name: The name of the variable to be plotted on the x-axis.
+                If empty, use the time variable.
+            time_unit: The unit to express the time.
+            save: Whether to save the figure.
+            show: Whether to show the figure.
+            file_path: The path of the file to save the figure.
+                The directory path and file format are deduced from it.
+                If empty,
+                save the file in the current directory,
+                with the output name as file name and PNG format.
+
+        Returns:
+            The figure.
+        """
+        if isinstance(output_names, str):
+            output_names = [output_names]
+
+        time_name = f"Time ({time_unit})"
+        if not abscissa_name:
+            abscissa_name = time_name
+
+        dataset = Dataset()
+        time_duration = TimeDuration(self.time[:, newaxis])
+        dataset.add_variable(time_name, time_duration.to(time_unit))
+        for name in set(output_names).union({abscissa_name}) - {time_name}:
+            dataset.add_variable(name, self.local_data[name][:, newaxis])
+
+        figure = Lines(dataset, output_names, abscissa_variable=abscissa_name)
+        figure.execute(save=save, show=show, file_path=file_path)
+        return figure